+++ src/ej08.py
@@ -1,13 +1,13 @@
 # Pedimos dos números de los tres
-num1 = int(input('Introduzca el primer número: '))
-num2 = int(input('Introduzca el segundo número: '))
+num1 = float(input('Introduzca el primer número: '))
+num2 = float(input('Introduzca el segundo número: '))
 
 # En esta operación sumamos el segundo número al primero y lo guardamos la suma en la primera
 # variable para dejar la segunda vacía
 num1 += num2  
 
 # Volvemos a asignar un valor a num2
-num2 = int(input('Introduzca el tercer número: '))
+num2 = float(input('Introduzca el tercer número: '))
 num1 += num2
 
 # Imprimimos el resultado
